requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
audioop-lts; python_version >= "3.13"
huggingface_hub>=0.26,<1.0
pydantic>=2.0,<2.10
gradio>=4.44.0,<5
transformers>=4.44.0,<5
torch>=2.0.0
Pillow>=10.0.0
